requirements.txt
playwright>=1.40.0
ollama>=0.1.0
pdfplumber>=0.10.0
pymupdf>=1.24.0
pyyaml>=6.0
rich>=13.0.0
click>=8.1.0
beautifulsoup4>=4.12.0
httpx>=0.25.0
lxml>=4.9.0
ttkbootstrap>=1.10.0
python-jobspy>=1.1.0
pandas>=2.0.0
imap_tools>=1.5.0
